# [Merged by Bors] - perf(VectorBundle/FiberwiseLinear): speed up